refactor(api): move update controller
in organizational-entities
