+++ api/lib/application/certification-centers/certification-center-controller.js
@@ -1,42 +1,9 @@
-import lodash from 'lodash';
-
-import { usecases as certificationConfigurationUsecases } from '../../../src/certification/configuration/domain/usecases/index.js';
-import * as certificationCenterForAdminSerializer from '../../../src/organizational-entities/infrastructure/serializers/jsonapi/certification-center/certification-center-for-admin.serializer.js';
 import * as divisionSerializer from '../../../src/prescription/campaign/infrastructure/serializers/jsonapi/division-serializer.js';
 import * as certificationCenterMembershipSerializer from '../../../src/shared/infrastructure/serializers/jsonapi/certification-center-membership.serializer.js';
 import { usecases as teamUsecases } from '../../../src/team/domain/usecases/index.js';
 import { usecases } from '../../domain/usecases/index.js';
-import { DomainTransaction } from '../../infrastructure/DomainTransaction.js';
 import * as sessionSummarySerializer from '../../infrastructure/serializers/jsonapi/session-summary-serializer.js';
 import * as studentCertificationSerializer from '../../infrastructure/serializers/jsonapi/student-certification-serializer.js';
-
-const { map } = lodash;
-
-const update = async function (request) {
-  const certificationCenterId = request.params.id;
-  const certificationCenterInformation = certificationCenterForAdminSerializer.deserialize(request.payload);
-  const complementaryCertificationIds = map(request.payload.data.relationships?.habilitations?.data, 'id');
-
-  const { updatedCertificationCenter, certificationCenterPilotFeatures } = await DomainTransaction.execute(
-    async () => {
-      const updatedCertificationCenter = await usecases.updateCertificationCenter({
-        certificationCenterId,
-        certificationCenterInformation,
-        complementaryCertificationIds,
-      });
-
-      const certificationCenterPilotFeatures = await certificationConfigurationUsecases.registerCenterPilotFeatures({
-        centerId: updatedCertificationCenter.id,
-        isV3Pilot: certificationCenterInformation.isV3Pilot,
-      });
-
-      return { updatedCertificationCenter, certificationCenterPilotFeatures };
-    },
-    { isolationLevel: 'repeatable read' },
-  );
-
-  return certificationCenterForAdminSerializer.serialize(updatedCertificationCenter, certificationCenterPilotFeatures);
-};
 
 const findPaginatedSessionSummaries = async function (request) {
   const certificationCenterId = request.params.id;
@@ -127,7 +94,6 @@ const certificationCenterController = {
   findPaginatedSessionSummaries,
   getDivisions,
   getStudents,
-  update,
   updateReferer,
 };
